Do not report back to GitHub with the PR buildeers (#193)
It seems that the PR builders can get confused if a pull requests
contains commits of another pull request when ussing the
`test-with-buildbots` label. For instance, this PR:

https://github.com/python/cpython/pull/20006

included messages that report to a commit that lives in another PR where
the label was added. To avoid this, exclude the PR builders from the
GitHub notification system.