Disable refleak builds on ware-win81-release (#319)
Something keeps killing this worker, and the current leading suspect is
one of the refleak builds.  It's also possible that something is going
wrong with Azure, as the worker does not recover after killing off
buildbot-worker.exe and all children.  Leaving it running for a few days
without the refleak builds should hopefully give some more information.
Azure seems to have completely dropped all possibility of creating a new
Windows 8.1 VM, so if it turns out to still be dying without the refleak
builds, it's probably going to wind up retired.

To keep some load on the worker, temporary unstable debug mode builders
are also added for both 32 and 64 bit builds.